backend/requirements/local.txt
django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
pillow==10.4.0
django-jazzmin==3.0.0
django-simple-history==3.7.0
django-filter==23.1
djangorestframework-simplejwt==5.3.1
drf-nested-forms==1.1.8
requests==2.32.3
django-treebeard==4.7
django-otp==1.3.0
celery==5.4.0
django-celery-beat==2.7.0
python-dateutil==2.9.0.post0
openpyxl==3.1.5
